move loaders into platform-specific monitor code
