--- monitors/elf.c
@@ -1,173 +1,0 @@
-
-
-static ssize_t pread_in_full(int fd, void *buf, size_t count, off_t offset)
-{
-    ssize_t total = 0;
-    char *p = buf;
-
-    if (count > SSIZE_MAX) {
-        errno = E2BIG;
-        return -1;
-    }
-
-    lseek(fd, 0, SEEK_SET);
-    while (count > 0) {
-        ssize_t nr;
-
-        lseek(fd, offset, SEEK_SET);
-        nr = read(fd, p, count);
-        if (nr == 0)
-            return total;
-        else if (nr == -1 && errno == EINTR)
-            continue;
-        else if (nr == -1)
-            return -1;
-
-        count -= nr;
-        total += nr;
-        p += nr;
-        offset += nr;
-    }
-
-    return total;
-}
-
-
-/*
- * Load code from elf file into *mem and return the elf entry point
- * and the last byte of the program when loaded into memory. This
- * accounts not only for the last loaded piece of code from the elf,
- * but also for the zeroed out pieces that are not loaded and sould be
- * reserved.
- *
- * Memory will look like this after the elf is loaded:
- *
- * *mem                    *p_entry                   *p_end
- *   |             |                    |                |
- *   |    ...      | .text .rodata      |   .data .bss   |
- *   |             |        code        |   00000000000  |
- *   |             |  [PROT_EXEC|READ]  |                |
- *
- */
-static void load_code(const char *file, uint8_t *mem,     /* IN */
-                      uint64_t *p_entry, uint64_t *p_end) /* OUT */
-{
-    int fd_kernel;
-    ssize_t numb;
-    size_t buflen;
-    Elf64_Off ph_off;
-    Elf64_Half ph_entsz;
-    Elf64_Half ph_cnt;
-    Elf64_Half ph_i;
-    Elf64_Phdr *phdr = NULL;
-    Elf64_Ehdr hdr;
-
-    /* elf entry point (on physical memory) */
-    *p_entry = 0;
-    /* highest byte of the program (on physical memory) */
-    *p_end = 0;
-
-    fd_kernel = open(file, O_RDONLY);
-    if (fd_kernel == -1)
-        goto out_error;
-
-    numb = pread_in_full(fd_kernel, &hdr, sizeof(Elf64_Ehdr), 0);
-    if (numb < 0)
-        goto out_error;
-    if (numb != sizeof(Elf64_Ehdr))
-        goto out_invalid;
-
-    /*
-     * Validate program is in ELF64 format:
-     * 1. EI_MAG fields 0, 1, 2, 3 spell ELFMAG('0x7f', 'E', 'L', 'F'),
-     * 2. File contains 64-bit objects,
-     * 3. Objects are Executable,
-     * 4. Target instruction set architecture is set to x86_64.
-     */
-    if (hdr.e_ident[EI_MAG0] != ELFMAG0
-            || hdr.e_ident[EI_MAG1] != ELFMAG1
-            || hdr.e_ident[EI_MAG2] != ELFMAG2
-            || hdr.e_ident[EI_MAG3] != ELFMAG3
-            || hdr.e_ident[EI_CLASS] != ELFCLASS64
-            || hdr.e_type != ET_EXEC
-            || hdr.e_machine != EM_X86_64)
-        goto out_invalid;
-
-    ph_off = hdr.e_phoff;
-    ph_entsz = hdr.e_phentsize;
-    ph_cnt = hdr.e_phnum;
-    buflen = ph_entsz * ph_cnt;
-
-    phdr = (Elf64_Phdr *)malloc(buflen);
-    if (!phdr)
-        goto out_error;
-    numb = pread_in_full(fd_kernel, phdr, buflen, ph_off);
-    if (numb < 0)
-        goto out_error;
-    if (numb != buflen)
-        goto out_invalid;
-
-    /*
-     * Load all segments with the LOAD directive from the elf file at offset
-     * p_offset, and copy that into p_addr in memory. The amount of bytes
-     * copied is p_filesz.  However, each segment should be given
-     * p_memsz aligned up to p_align bytes on memory.
-     */
-    for (ph_i = 0; ph_i < ph_cnt; ph_i++) {
-        uint8_t *daddr;
-        uint64_t _end;
-        size_t offset = phdr[ph_i].p_offset;
-        size_t filesz = phdr[ph_i].p_filesz;
-        size_t memsz = phdr[ph_i].p_memsz;
-        uint64_t paddr = phdr[ph_i].p_paddr;
-        uint64_t align = phdr[ph_i].p_align;
-        uint64_t result;
-
-        if (phdr[ph_i].p_type != PT_LOAD)
-            continue;
-
-        if ((paddr >= GUEST_SIZE) || add_overflow(paddr, filesz, result)
-                || (result >= GUEST_SIZE))
-            goto out_invalid;
-        if (add_overflow(paddr, memsz, result) || (result >= GUEST_SIZE))
-            goto out_invalid;
-        /*
-         * Verify that align is a non-zero power of 2 and safely compute
-         * ((_end + (align - 1)) & -align).
-         */
-        if (align > 0 && (align & (align - 1)) == 0) {
-            if (add_overflow(result, (align - 1), _end))
-                goto out_invalid;
-            _end = _end & -align;
-        } else {
-            _end = result;
-        }
-        if (_end > *p_end)
-            *p_end = _end;
-
-        daddr = mem + paddr;
-        numb = pread_in_full(fd_kernel, daddr, filesz, offset);
-        if (numb < 0)
-            goto out_error;
-        if (numb != filesz)
-            goto out_invalid;
-        memset(daddr + filesz, 0, memsz - filesz);
-
-        /* Write-protect the executable segment */
-        if (phdr[ph_i].p_flags & PF_X) {
-            if (mprotect(daddr, _end - paddr, PROT_EXEC | PROT_READ) == -1)
-                goto out_error;
-        }
-    }
-
-    free(phdr);
-    close(fd_kernel);
-    *p_entry = hdr.e_entry;
-    return;
-
-out_error:
-    err(1, "%s", file);
-
-out_invalid:
-    errx(1, "%s: Exec format error", file);
-}
